# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management and Audit System API",
        "description": "A comprehensive API for inventory management with audit logging, role-based access control, and real-time notifications",
        "contact": {
            "name": "API Support",
            "email": "admin@inventory-system.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Inventory Management API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/audit-logs": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "List audit logs",
                "description": "Returns a paginated list of audit logs with filtering options",
                "operationId": "auditLogIndex",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by audit log type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "item_created",
                                "item_updated",
                                "item_deleted",
                                "stock_adjusted"
                            ]
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "inventory_item_id",
                        "in": "query",
                        "description": "Filter by inventory item ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AuditLogResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuditLogResource": {
                "title": "Audit Log Resource",
                "description": "Audit log resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "item_created",
                            "item_updated",
                            "item_deleted",
                            "stock_adjusted"
                        ],
                        "example": "item_created"
                    },
                    "description": {
                        "type": "string",
                        "example": "Created inventory item: Laptop XPS 15"
                    },
                    "user_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "warehouse": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Main Warehouse"
                            },
                            "location": {
                                "type": "string",
                                "example": "New York"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object"
                    },
                    "inventory_item": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Laptop XPS 15"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-09T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}